ref: move entities to db folder
