+++ src/db/entities/base-repository.ts
